feat(agent-tools): add AgentDialog, AgentPanel, and McpServerPanel components for managing agents and MCP servers
