+++ sites/playground/web/src/components/tab-components/tool-panels/AgentDialog.vue
